Split domain classes into entity and enum packages
and some typos
