+++ backend/src/main/java/de/op/atom/foods/domain/entity/Recipe.java
@@ -1,4 +1,4 @@
-package de.op.atom.foods.domain;
+package de.op.atom.foods.domain.entity;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
